refactor(scheduler): extract monitor wait read model
